Move ReadTransformedResourceOperation into subsystem-test, as it is only used from there
